+++ demos/07-voronoi-and-delaunay/index.js
@@ -91,38 +91,25 @@
 				   Math.random()*pb.canvasSize.height*0.5 - pb.canvasSize.height/2*0.5
 				 );
 	    };
-	    var randomPoint = randomVertex;
+	    // var randomPoint = randomVertex;
 
 	    
 	    // +---------------------------------------------------------------------------------
 	    // | A global config that's attached to the dat.gui control interface.
 	    // +-------------------------------
 	    var config = PlotBoilerplate.utils.safeMergeByKeys( {
-		fillTriangles         : true,
-		trianglesInGreyscale  : false,
 		makeVoronoiDiagram    : true,
-		//fillAlphaOnly         : false,
 		drawPoints            : true,
-		drawEdges             : true,
+		drawTriangles         : true,
 		drawCircumCircles     : false,
-		//drawQuadraticCurves   : false,
 		drawCubicCurves       : false,
 		voronoiCubicThreshold : 1.0,
-		optimizeGaps          : false,
 		pointCount            : 25,
-		triangulate           : true,
-		autoUpdateOnChange    : true,
 		backgroundColor       : '#ffffff',
 		rebuild               : function() { rebuild(); },
-		//loadImage             : function() { var elem = document.getElementById('file'); elem.setAttribute('data-type','image-upload'); triggerClickEvent(elem); },
-		//fitImage              : true,
-		clear                 : function() { pointList = []; triangles = []; voronoiDiagram = []; redraw(); },
 		randomize             : function() { randomPoints(true,false,false); trianglesPointCount = -1; rebuild(); },
-		fullCover             : function() { randomPoints(true,true,false); trianglesPointCount = -1; pb.rebuild(); },
-		fullCoverExtended     : function() { randomPoints(true,true,false); trianglesPointCount = -1; rebuild() }
-		//exportSVG             : function() { exportSVG(); },
-		//exportPointset        : function() { exportPointset(); },
-		//importPointset        : function() { var elem = document.getElementById('file'); elem.setAttribute('data-type','pointset-upload'); triggerClickEvent(elem); } 
+		fullCover             : function() { randomPoints(true,true,false); trianglesPointCount = -1; rebuild(); },
+		//fullCoverExtended     : function() { randomPoints(true,true,false); trianglesPointCount = -1; rebuild() }
 	    }, GUP );
 
 
@@ -138,7 +125,7 @@
 	    // | Adds a random point to the point list. Needed for initialization.
 	    // +-------------------------------
 	    var addRandomPoint = function() {
-		addVertex( randomPoint() );
+		addVertex( randomVertex() );
 		
 	    };
 
@@ -158,35 +145,10 @@
 	    };
 
 	    // +---------------------------------------------------------------------------------
-	    // | A negative-numbers friendly max function (determines the absolute max and
-	    // | preserves the signum).
-	    // +-------------------------------
-	    var absMax = function(value,max) {
-		if( Math.abs(value) < max ) return Math.sign(value)*max;
-		return value;
-	    };
-
-
-	    // +---------------------------------------------------------------------------------
 	    // | Generates a random int value between 0 and max (both inclusive).
 	    // +-------------------------------
 	    var randomInt = function(max) {
 		return Math.round( Math.random()*max );
-	    };
-
-	    // +---------------------------------------------------------------------------------
-	    // | Generates a random color object.
-	    // +-------------------------------
-	    var randomColor = function() {
-		return Color.makeRGB( randomInt(255), randomInt(255), randomInt(255) );
-	    };
-
-	    // +---------------------------------------------------------------------------------
-	    // | Generates a random color object with r=g=b.
-	    // +-------------------------------
-	    var randomGreyscale = function() {
-		var v = 32 + randomInt(255-32);
-		return Color.makeRGB( v, v, v );
 	    };
 	    
 	    // +---------------------------------------------------------------------------------
@@ -198,20 +160,14 @@
 		pb.draw.line( t.c, t.a, color );
 	    };
 	    
-	    // +---------------------------------------------------------------------------------
-	    // | Get average color in triangle.
-	    // | @param imageBuffer:canvas
-	    // +-------------------------------
-	    var getAverageColorInTriangle = function( imageBuffer, tri ) {
-		return randomColor();
-	    };
 	    
 	    // +---------------------------------------------------------------------------------
 	    // | The re-drawing function.
 	    // +-------------------------------
 	    var redraw = function() {
 		// Draw triangles
-		drawTriangles();
+		if( config.drawTriangles )
+		    drawTriangles();
 
 		// Draw circumcircles
 		if( config.drawCircumCircles )
@@ -220,10 +176,6 @@
 		// Draw voronoi diagram?
 		if( config.makeVoronoiDiagram )
 		    drawVoronoiDiagram();
-		
-		// Draw quadratic curves?
-		if( config.drawQuadraticCurves )
-		    drawQuadraticBezierVoronoi();
 
 		// Draw cubic curves
 		if( config.drawCubicCurves )
@@ -272,60 +224,11 @@
 		    if( cell.isOpen() || cell.triangles.length < 3 )
 			continue;
 		    var cbezier = new Polygon(cell.toPathArray(),cell.isOpen()).toCubicBezierData( config.voronoiCubicThreshold );
-		    pb.draw.cubicBezierPath( cbezier, '#0048e8' );
+		    pb.draw.cubicBezierPath( cbezier, '#0088e8' );
 		}
 		
 	    }; // END drawCubicBezierVoronoi
 	    
-
-	    // +---------------------------------------------------------------------------------
-	    // | Handle a dropped JSON file (pointset data).
-	    // |
-	    // | The required data format is a JSON string representing array:{x,y}
-	    // |
-	    // | Example:
-	    // |  '[ { x : 5, y : 6 }, { x : 100, y : 20 }, { x : 23.5, y : -10 } ]'
-	    // +-------------------------------
-	    /*var handlePointset = function(e) {
-		var reader = new FileReader();
-		reader.onload = function(event){
-		    var json = event.target.result;
-		    try {
-			var pointset = JSON.parse(json);		    
-			pointList = [];
-			for( var i in pointset ) {
-			    var tuple = pointset[i];
-			    pointList.push( new Vertex(tuple.x,tuple.y) );
-			}
-			if( config.autoUpdateOnChange ) rebuild();
-			else                            redraw();
-		    } catch( e ) {
-			console.log( JSON.stringify(pointset) );
-			console.log( e );		    
-		    }
-		}
-		reader.readAsText(e.target.files[0]);     
-	    }*/
-
-	    
-	    // +---------------------------------------------------------------------------------
-	    // | Decide which file type should be handled:
-	    // |  - image for the background or
-	    // |  - JSON for the point set)
-	    // +-------------------------------
-	    /*var handleFile = function(e) {
-		var type = document.getElementById('file').getAttribute('data-type');
-		if( type == 'image-upload' ) {
-		    handleImage(e);
-		} else if( type == 'pointset-upload' ) {
-		    handlePointset(e);
-		} else {
-		    console.warn('Unrecognized upload type: ' + type );
-		}   
-	    }
-	    document.getElementById( 'file' ).addEventListener('change', handleFile );
-	    */
-
 	    
 	    // +---------------------------------------------------------------------------------
 	    // | The rebuild function just evaluates the input and
@@ -335,13 +238,13 @@
 	    var rebuild = function() {
 		// Only re-triangulate if the point list changed.
 		var draw = true;
-		if( (config.triangulate || config.makeVoronoiDiagram) ) // && trianglesPointCount != pointList.length )
-		    triangulate();
-		if( config.makeVoronoiDiagram )
+		//if( (config.triangulate || config.makeVoronoiDiagram) ) // && trianglesPointCount != pointList.length )
+		triangulate();
+		if( config.makeVoronoiDiagram || config.drawCubicCurves )
 		    draw = makeVoronoiDiagram();
 
 		if( draw )
-		    redraw();
+		    pb.redraw();
 	    };
 
 	    
@@ -349,23 +252,8 @@
 	    // | Make the triangulation (Delaunay).
 	    // +-------------------------------
 	    var triangulate = function() {
-		//console.log( window.Delaunay );
 		var delau = new Delaunay( pointList, {} );
 		triangles  = delau.triangulate();
-
-		// Optimize triangles?
-		if( config.optimizeGaps > 0 ) {
-		    for( var i in triangles ) {
-			var tri = triangles[i];
-			var circumCircle = tri.getCircumcircle(); // { center:Vertex, radius:Number }
-			var scaleFactor = (circumCircle.radius+0.1) / circumCircle.radius;
-			
-			tri.a = tri.a.clone().scale( scaleFactor, circumCircle.center );
-			tri.b = tri.b.clone().scale( scaleFactor, circumCircle.center );
-			tri.c = tri.c.clone().scale( scaleFactor, circumCircle.center );
-		    }
-		}
-
 		trianglesPointCount = pointList.length;
 		voronoiDiagram = [];
 		redraw();
@@ -403,42 +291,46 @@
 	    // +---------------------------------------------------------------------------------
 	    // | Add n random points.
 	    // +-------------------------------
-	    var randomPoints = function( clear, fullCover, doRedraw ) {
-		if( clear )
+	    var randomPoints = function( clear, fullCover, doRebuild ) {
+		if( clear ) {
+		    for( var i in pointList )
+			pb.remove( pointList[i], false );
 		    pointList = [];
+		    console.log( pointList, pb.drawables );
+		}
 		// Generate random points on image border?
 		if( fullCover ) {
 		    var remainingPoints = config.pointCount-pointList.length;
 		    var borderPoints    = Math.sqrt(remainingPoints);
-		    var ratio           = canvasSize.height/canvasSize.width;
+		    var ratio           = pb.canvasSize.height/pb.canvasSize.width;
 		    var hCount          = Math.round( (borderPoints/2)*ratio );
 		    var vCount          = (borderPoints/2)-hCount;
 		    
 		    while( vCount > 0 ) {
-			pointList.push( new Vertex(0, randomInt(canvasSize.height)) );
-			pointList.push( new Vertex(canvasSize.width, randomInt(canvasSize.height)) );		    
+			addVertex( new Vertex(-pb.canvasSize.width/2, randomInt(pb.canvasSize.height/2)-pb.canvasSize.height/2) );
+			addVertex( new Vertex(pb.canvasSize.width/2, randomInt(pb.canvasSize.height/2)-pb.canvasSize.height/2) );		    
 			vCount--;
 		    }
 		    
 		    while( hCount > 0 ) {
-			pointList.push( new Vertex(randomInt(canvasSize.width),0) );
-			pointList.push( new Vertex(randomInt(canvasSize.width),canvasSize.height) );		    
+			addVertex( new Vertex(randomInt(pb.canvasSize.width/2)-pb.canvasSize.width/2,0) );
+			addVertex( new Vertex(randomInt(pb.canvasSize.width/2)-pb.canvasSize.width/2,pb.canvasSize.height/2) );
 			hCount--;
 		    }
 
 		    // Additionally add 4 points to the corners
-		    pointList.push( new Vertex(0,0) );
-		    pointList.push( new Vertex(canvasSize.width,0) );
-		    pointList.push( new Vertex(canvasSize.width,canvasSize.height) );
-		    pointList.push( new Vertex(0,canvasSize.height) );	
+		    addVertex( new Vertex(0,0) );
+		    addVertex( new Vertex(pb.canvasSize.width/2,0) );
+		    addVertex( new Vertex(pb.canvasSize.width/2,pb.canvasSize.height/2) );
+		    addVertex( new Vertex(0,pb.canvasSize.height/2) );	
 		}
 		
 		// Generate random points.
 		for( var i = pointList.length; i < config.pointCount; i++ ) {
 		    addRandomPoint();
 		}
-		if( doRedraw )
-		    redraw();
+		if( doRebuild )
+		    rebuild();
 	    };
 
 	    // +---------------------------------------------------------------------------------
@@ -446,11 +338,13 @@
 	    // +-------------------------------
 	    var updatePointCount = function() {
 		if( config.pointCount > pointList.length )
-		    randomPoints(false,false,true); // Do not clear ; no full cover ; do redraw
+		    randomPoints(false,false,true); // Do not clear ; no full cover ; do rebuild
 		else if( config.pointCount < pointList.length ) {
 		    // Remove n-m points
+		    for( var i = config.pointCount; i < pointList.length; i++ )
+			pb.remove( pointList[i] );
 		    pointList = pointList.slice( 0, config.pointCount );
-		    redraw();
+		    rebuild();
 		}
 		
 	    };
@@ -479,30 +373,17 @@
 		f0.add(config, 'pointCount').min(3).max(200).onChange( function() { config.pointCount = Math.round(config.pointCount); updatePointCount(); } ).title("The total number of points.");
 		f0.add(config, 'randomize').name('Randomize').title("Randomize the point set.");
 		f0.add(config, 'fullCover').name('Full Cover').title("Randomize the point set with full canvas coverage.");
-		f0.add(config, 'clear').name('Clear all').title("Clear all.");
-		f0.add(config, 'drawPoints').onChange( redraw ).title("If checked the points will be drawn.");
+		//f0.add(config, 'drawPoints').onChange( function() { pb.redraw() } ).title("If checked the points will be drawn.");
 		f0.open();
 		
 		var f1 = gui.addFolder('Delaunay');
-		f1.add(config, 'triangulate').onChange( rebuild ).title("Triangulate the point set?");
-		f1.add(config, 'fillTriangles').onChange( redraw ).title("If selected the triangles will be filled."); 
-		//f1.add(config, 'fillAlphaOnly').onChange( redraw ).title("Only the alpha channel from the image will be applied.");
-		//f1.add(config, 'trianglesInGreyscale').onChange( reassignTriangleColors ).title("If selected the triangles will be filled in greyscale only.");
-		f1.add(config, 'drawEdges').onChange( redraw ).title("If checked the triangle edges will be drawn.");
-		f1.add(config, 'drawCircumCircles').onChange( pb.redraw ).title("If checked the triangles circumcircles will be drawn.");
-		//f1.add(config, 'optimizeGaps').onChange( pb.rebuild ).title("If checked the triangles are scaled by 0.15 pixels to optimize gaps.");
+		f1.add(config, 'drawTriangles').onChange( function() { pb.redraw() } ).title("If checked the triangle edges will be drawn.");
+		f1.add(config, 'drawCircumCircles').onChange( function() { pb.redraw() } ).title("If checked the triangles circumcircles will be drawn.");
 
 		var f2 = gui.addFolder('Voronoi');
 		f2.add(config, 'makeVoronoiDiagram').onChange( rebuild ).title("Make voronoi diagram from the triangle set.");
-		f2.add(config, 'drawCubicCurves').onChange( pb.redraw ).title("If checked the Voronoi's cubic curves will be drawn.");
-		f2.add(config, 'voronoiCubicThreshold').min(0.0).max(1.0).onChange( pb.redraw ).title("(Experimental) Specifiy the cubic coefficients.");
-		
-		/*
-		  var f4 = gui.addFolder('Import & Export');
-		  f4.add(config, 'exportSVG').name('Export SVG').title("Export the current triangulation as a vector image.");
-		  f4.add(config, 'exportPointset').name('Export point set').title("Export the point set as JSON.");
-		  f4.add(config, 'importPointset').name('Import point set').title("Import the point set from JSON.");	    
-		*/
+		f2.add(config, 'drawCubicCurves').onChange( rebuild ).title("If checked the Voronoi's cubic curves will be drawn.");
+		f2.add(config, 'voronoiCubicThreshold').min(0.0).max(1.0).onChange( function() { pb.redraw() } ).title("(Experimental) Specifiy the cubic coefficients.");
 	    }